requirements.txt
piazza_api
requests
redis
python-dateutil
html-slacker
psycopg2-binary